Move the port parsing code into the `echoip` library.
We'll need to parse port numbers from the command line in both the
client and server `echoip` commands, so it makes sense to move the
`clap` argument generation into the `echoip` library shared by both
commands. As a bonus, we'll also be able to do a better job of testing
it.