Migrate `Config` to `ClientConfig`.
Since the rest of the client CLI parsing code has been moved over to
`config.rs`, we move the `Config` struct over as well. Of course, it
needs to be renamed to `ClientConfig`, as there will eventually be a
`ServerConfig` as well.

This change follows the pattern of leaving the minimum code in the
`echoip.rs` binary stub.