Add server loop to library crate.
We create a server loop, which listens on a specified UDP socket and
responds to the sender with the sender's IP address whenever it receives
a packet. The loop never exits.

This is essentially the entirety of the server program, less the command
line parsing. Tests are included for each function called by the server
loop, but not for the loop itself, as this would result in the tests
never finishing.