Add client functionallity to library crate.
We create a `client()` function, which sends a UDP packet to a server
running the `echoipd` daemon, and then listens for a reply from the
server containing the client's public IP address.

This is essentially the entirety of the client program, less the command
line parsing. Tests are included for each function called by the
`client()` function, but not for the function itself, as this would
potentially terminate the tests prior to completion.