Remove `client::decode_addr()`.
Since we now check for the size of the returned data in `recv_from()`,
we can return a fixed size array from this function instead of a `Vec`.
This means that conversion into a `Ipv4Addr` is trivial, and can be
performed in the main `client()` function instead of in `decode_addr()`.
As a result, we remove both `decode_addr()` and its test code.